pyproject.toml
[tool.poetry]
name = "navis-map-hack"
version = "0.1.0"
description = ""
authors = ["convoloot <joe.chan@biarri.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.3"
dash = "^1.11.0"
airtable-python-wrapper = "^0.13.0"
python-dotenv = "^0.13.0"
geocoder = "^1.38.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
app = 'entry:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

